% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKmodel-Package.R
\docType{data}
\name{ajk_Haber_table3}
\alias{ajk_Haber_table3}
\title{Houshold infeection data table
Haber et al. (1988)) Table3}
\format{A dataframe}
\description{
Houshold infeection data table
Haber et al. (1988)) Table3
}
\keyword{datasets}
