% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKmodel-Package.R
\docType{data}
\name{ajk_Longini_table3crowded}
\alias{ajk_Longini_table3crowded}
\title{Houshold infeection data table
#' Longini and Koopman (1982) Table3}
\format{A dataframe}
\description{
Houshold infeection data table
#' Longini and Koopman (1982) Table3
}
\keyword{datasets}
