% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathModel.R
\name{estmjk}
\alias{estmjk}
\title{estimating probability}
\usage{
estmjk(B, Q)
}
\arguments{
\item{B}{a probability of infection from cumunity}

\item{Q}{a probability of infection from household}
}
\value{
A probability table
}
\description{
estimating probability
}
