% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKmodel-Package.R
\docType{package}
\name{LKmodel}
\alias{LKmodel}
\alias{LKmodel-package}
\title{LKmodel: A package for household infectious model.}
\description{
LKmodel: A package for household infectious model.
}
\references{
{
Longini IM, Koopman JS (1982) Household and community transmission parameters from final distributions of infections in households.
\emph{Biometrics} \bold{38}, 115-126
}
}
