% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathModel.R
\name{LKmodel.test}
\alias{LKmodel.test}
\title{LKmodel.test}
\usage{
LKmodel.test(x = NULL, y = NULL, method = "SC")
}
\arguments{
\item{x}{a dataset of household}

\item{y}{another dataset of household}

\item{method}{test method}
}
\value{
test results
}
\description{
LKmodel.test
}
\examples{
fm <- LKmodel.test(x=ajk_Longini_table1)
summary(fm)
}
