% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathModel.R
\name{summary.default}
\alias{summary.default}
\title{Suumary default method}
\usage{
\method{summary}{default}(object, ...)
}
\arguments{
\item{object}{LKmodel.test obeject}

\item{...}{other arguments}
}
\description{
Suumary default method
}
