% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathModel.R
\name{summary.LKmodel.test}
\alias{summary.LKmodel.test}
\title{Suumary LK-model.test method}
\usage{
\method{summary}{LKmodel.test}(object, ...)
}
\arguments{
\item{object}{LKmodel.test obeject}

\item{...}{other arguments}
}
\description{
Suumary LK-model.test method
}
