% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathModel.R
\name{summary}
\alias{summary}
\title{summary: summary function}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{LKmodel.test obeject}

\item{...}{other arguments}
}
\description{
A details of summary
}
